Update paths for authentication note info
