+++ crates/block-producer/src/batch_builder/batch.rs
@@ -3,7 +3,7 @@ use std::{
     mem,
 };
 
-use miden_node_store::state::NoteAuthenticationInfo;
+use miden_node_proto::domain::notes::NoteAuthenticationInfo;
 use miden_objects::{
     accounts::{delta::AccountUpdateDetails, AccountId},
     batches::BatchNoteTree,
