+++ crates/proto/src/domain/notes.rs
@@ -1,13 +1,19 @@
+use std::collections::{BTreeMap, BTreeSet};
+
 use miden_objects::{
     notes::{NoteExecutionHint, NoteId, NoteInclusionProof, NoteMetadata, NoteTag, NoteType},
     Digest, Felt,
 };
 
 use crate::{
+    convert,
+    domain::blocks::BlockInclusionProof,
     errors::{ConversionError, MissingFieldHelper},
     generated::note::{
+        NoteAuthenticationInfo as NoteAuthenticationInfoProto,
         NoteInclusionInBlockProof as NoteInclusionInBlockProofPb, NoteMetadata as NoteMetadataPb,
     },
+    try_convert,
 };
 
 impl TryFrom<NoteMetadataPb> for NoteMetadata {
@@ -82,5 +88,43 @@ impl TryFrom<&NoteInclusionInBlockProofPb> for (NoteId, NoteInclusionProof) {
                     .try_into()?,
             )?,
         ))
+    }
+}
+
+#[derive(Clone, Default, Debug)]
+pub struct NoteAuthenticationInfo {
+    pub block_proofs: Vec<BlockInclusionProof>,
+    pub note_proofs: BTreeMap<NoteId, NoteInclusionProof>,
+}
+
+impl NoteAuthenticationInfo {
+    pub fn contains_note(&self, note: &NoteId) -> bool {
+        self.note_proofs.contains_key(note)
+    }
+
+    pub fn note_ids(&self) -> BTreeSet<NoteId> {
+        self.note_proofs.keys().copied().collect()
+    }
+}
+
+impl From<NoteAuthenticationInfo> for NoteAuthenticationInfoProto {
+    fn from(value: NoteAuthenticationInfo) -> Self {
+        Self {
+            note_proofs: convert(&value.note_proofs),
+            block_proofs: convert(value.block_proofs),
+        }
+    }
+}
+
+impl TryFrom<NoteAuthenticationInfoProto> for NoteAuthenticationInfo {
+    type Error = ConversionError;
+
+    fn try_from(value: NoteAuthenticationInfoProto) -> Result<Self, Self::Error> {
+        let result = Self {
+            block_proofs: try_convert(value.block_proofs)?,
+            note_proofs: try_convert(&value.note_proofs)?,
+        };
+
+        Ok(result)
     }
 }
